Improvements
Higly improved logger, it can now log tables and empty lines as well as multiple lines string.

Changed the config to Yaml but I dunno if it's a good idea, we can read but not save. However it makes it look fancy with comments for the user (which is very important)